Kubernetes Developer's Summit Schedule
The  schedule for the Dev Summit is posted. Barring extraneous circumstances, the discussions listed are the discussions that will be held during the summit. These topics came directly from the top 12 choices of the CIVS poll, so thank you all for voting. 

We may have to shuffle *the order* of the discussions, if any discussion facilitators have scheduling conflicts, but otherwise this should be the final schedule!

The formatting will change for any day-of handouts, but we've gone with the simplest format available in order to push the schedule to the community quickly.