Update SIG OpenStack Meeting times for 2017
As we closed out 2016 we ran a poll to select new meeting times for the
coming year. As discussed in the last meeting there was no clear winner
in the results but there were a handful of slots in particular that came
up as "least bad".

In the hope of encouraging participation from those who have not been able
to attend in the past we are going to trial running a 1500 UTC meeting time,
alternating with our traditional 2100 UTC meeting time. Meetings remain
only every second week for now.