Migrate the remaining content out of the wiki
I anticipate much of this is obsolete or outdated.  I outright deleted
content for which I knew this to be the case.  What remains is content
I don't have direct knowledge of or control over.  If you would rather
I just delete it than dump in your SIG's directory, please speak up