contrib guide: tighten some language
The difference between the two links for communications around questions
and troubleshooting was not clear to me, and the wording made them
seem somewhat redundant.  They're really two separate topic links,
so I've changed the wording to hopefully make it more clear that one has
general communications information and once is actually a
troubleshooting guide.

Signed-off-by: Tim Pepper <tpepper@vmware.com>