remove contributors/devel/go-code.md
The go-code.md file is mostly redundant relative to
contributors/devel/development.md and godep.md, the
contributors/guide/coding-conventions.md document and the various
now standard and easily searchable k8s external sources of information
about the Go language.

The redundant links are simply dropped.

The moovweg/gvm link is dropped as the project appears stale with no
releases in years, many open issues, no commits in a year and shows a build
failure in its CI badge.

The emptied go-code.md file is retained, leaving a pointer to the
new location in its place.

Signed-off-by: Tim Pepper <tpepper@vmware.com>