Change KEP state from accepted to draft
This PR is a proposal to make the initial state of a KEP more clear.
There have been misunderstandings on the means because of the way
accepted has been used elsewhere on the project.

This change moves the term to draft which is used buy other projects
such as the IETF standards track.