simplify the cherry pick process documentation
The cherry-pick instruction guide was quite out of date.  This
commit dramatically simplifies the doc.  Also, based on watching a
new user attempt a cherry pick (a common occurance around release
time with the revolving staffing of the release team), a few
clarification and setup issues are addressed around the
hack/cherry_pick_pull.sh script pre-requisites.

This is an interim fix as the doc will need updated again to capture
the pending prow based method when that becomes preferred and the
file should be merged with the various other branch and release
maintenance documents that are out there.

Signed-off-by: Tim Pepper <tpepper@vmware.com>