add info on joining Google Groups
As evidenced by a post to kubernetes-dev asking to be added to the list,
the current info on mailing lists may not be completely clear to
readers.  This commit makes the wording more explicit and includes a
link to Google Groups' info on how to find and join a group.

A broken link to the sig-list.md is also corrected.

Signed-off-by: Tim Pepper <tpepper@vmware.com>