*: Assorted internal-reference cleanups
This is mostly replacing
https://github.com/kubernetes/community/blob/master/... links with
relative links, because relative links are easier to maintain.  For
example, if you rename a.md to b.md in a feature branch, updating an
/a.md to /b.md target will work immediately (e.g. during pull-request
review), while updating
https://github.com/kubernetes/community/blob/master/a.md to
https://github.com/kubernetes/community/blob/master/b.md will only
work after your branch lands the rename in master.

I'm not sure why the #L1454 fragment was part of the old
sig-governance template sigs.yaml link, but it's been there since the
template landed in 68658bff (Provide short template for SIG
governance, 2018-02-23, #1830).  I've removed it in this commit.

The Kubernetes Charter README link in the sig-governance template
could just be:

  [Kubernetes Charter README]: README.md

but I've based it on the project root and back to make life easier
on new SIGs which copy/paste the file to a new location in this
repository.

I personally prefer targets that use ../ for referencing other
directories within a project, because those also work with generic
Markdown tools (e.g. Emacs' Markdown mode).  I've gone with /-rooted
targets in this commit as requested in
/sig-contributor-experience/markdown-link-style-guide.md.

I've also chased a few files and anchors as they've moved around,
e.g. the OBSOLETE_templates.md rename from 18958f82 (addressing Brians
comments, 2017-09-06, #1010) and the "and maintenance" header addition
from f471f6b6 (Add new contact info for Paris and me, 2017-10-17,

This commit isn't complete, but it's a step in the right direction and
all I had time for tonight ;).

The sig-list.md changes were generated automatically from the
generator/list.tmpl changes with:

  $ make generate