hack/verify-generated-docs.sh: clarify error output
When the hack/verify-generated-docs.sh job fails because a generated
file was edited instead of the sigs.yaml file, it is less than clear
from the error message that the user still needs to manually commit the
generated docs in addition to changing the sigs.yaml and validating that
the docs do build.  A small additional bit of guidance is appended to
the existing hint in the test error output.

Signed-off-by: Tim Pepper <tpepper@vmware.com>