contributors/devel: request better godep vendoring commit messages
On the release team I've seen a number of issues go by this year
where vendor code was updated and later implicated in issue triage
and bisection, but the original commits didn't sufficiently document
the change.  Reverting the update then reintroduced a previously
fixed issue, which was insufficiently documented in the changelog.  The
subsequent churn led to a lot of work by multiple people to resolve.

Terse godep vendor code update commit messages seem to be the norm and
we should aim higher.

Signed-off-by: Tim Pepper <tpepper@vmware.com>