steering: governance: define chair removal process
As outlined by [Brian Grant](https://github.com/kubernetes/community/pull/2040#discussion_r189472207)
there isn't a process for a SIG Chair to be removed. I think the best
course is to suggest an unresponsive chair be encouraged to step down
and if that fails a consensus is reached amongst existing chairs and
subproject owners.