Update SIG Autoscaling in sigs.yaml
This updates the mission statement for SIG Autoscaling to correspond to
the scope of SIG Autoscaling in the charter.

It also ensures that SIG Autoscaling has the right subprojects listed (cluster
autoscaler, vertical pod autoscaler, horizontal-pod-autoscaler, and
scale-client, cluster proportional autoscalers, addon-resizer), moving the
metrics project to SIG instrumentation where it belongs.