contrib & devel guide: pull in issue doc from sig-release/ephemera
The sig-release repo has a set of old documentation looking for a
better home.  The issues.md document there describes how a developer
targets an issue (or PR) to a milestone release.  This logically
fits better in the community/contributors/devel/ guide, so this commit
brings it in and a parallel commit will remove it from the other repo.

Signed-off-by: Tim Pepper <tpepper@vmware.com>