Rewrite 2018 voting guideline
This points to kubernetes/steering as the source of truth
for eligibility requirements for voting and candidacy

I also word wrapped the 2018 voting guide to 80 chars, which
makes for an awful diff, so I'll explain what I changed as
notes in the PR

I swear I didn't sneak in a "all votes secretly go to Aaron"
policy. Promise.  Really.