devel guide: pull in issue doc from sig-release/ephemera
The sig-release repo has a set of old documentation looking for a
better home.  The issues.md document there describes how a developer
targets an issue (or PR) to a milestone release.  This logically
fits better in the community/contributors/devel/ guide, so this commit
brings it in and a parallel commit will remove it from the other repo.

With it comes two pictures of what a release cycle and the overall
release lifecycle look like.

The issues.md content is modified a bit to add more information
about the release process overall and remove outdated information
around the bot-driven workflow for targeting work to a milestone, and is
named release.md here.

Signed-off-by: Tim Pepper <tpepper@vmware.com>