release docs: clarify tldr table html
I incorrectly had indicated a milestone was required in each of the four
major phases of the release cycle, where it is not in the "normal dev"
phase.  Not sure how we got that through review.  Part of it may have
been the html table format.  To help with that, in addition to removing
the errant indication of milestone as required in that early period,
I've also added html comments in the table so the reader of the raw code
can more clearly see which section is which.

Signed-off-by: Tim Pepper <tpepper@vmware.com>