Redo devstats as a subproject
Subprojects are supposed to be based on OWNERS files, even if they
are purely non-technical with no code they should still have docs
that explain what they are and how they operate.

I added a skeleton readme for devstats, and an OWNERS file. I added
jberkus and myself as approvers in addition to Phillels as I feel
we have done substantial work with devstats, if nothing else through
graph of the week presentations at community.