wg-lts: add
As per October 2018 discussion in SIG Release, k-dev mailing list,
etc., here's our proposal for the working group around evaluating
our support stance and making proposals for improvements.

While not required for a working group, this commit proposes an
initial charter for the WG LTS as this one has a notable potential
to at best bikeshed and at worst get mired in contention.  Documenting
scope, goals, deliverables seems like a reasonable way to aim for
a collaborative analysis of where we are and constructive suggestions
toward improvement.

Signed-off-by: Tim Pepper <tpepper@vmware.com>