Specify need for using py2 with Bazel
The default python on my system is python3. This caused me to run into
https://github.com/bazelbuild/rules_docker/issues/454 when running
`bazel build //...`. Update the documentation to indicate that
`/usr/bin/env python` must resolve to python2 in order for all of the
Bazel commands included in the documentation to work.

We can remove this comment after Bazel better supports using python3 as
the default system python. It appears from
https://github.com/bazelbuild/rules_docker/issues/580 that the
`rules_docker` folks hope to merge the fix in mid to late Q1 2019.