Add explicit links to release-engineering in sig-release
It's confusing having two links for the release-engineering subproject
without any description of each. The k/release repo houses all of the
existing tooling, such as anago, while the k/sig-release repo houses the
procedures and processes. Add these links explicitly to the description
so that newcomers have a clearer idea of where to get started.