github-management: clarify sponsorship criteria
An approver/reviewer in @kubernetes, may sponsor someone for the
kubernetes org or any of the related organizations
(kubernetes-sigs, kubernetes-csi etc); as long as it's a project
they're involved with.

However, this does not work the other way around.
A sponsor that is only an approver/reviewer in kubernetes-sigs cannot
sponsor someone for membership in the kubernetes org.
They are scoped just to the org they're associated with.