Add stakeholder SIGs for wg-policy
Add SIG Architecture, SIG Auth, SIG Network, SIG Node, SIG Multicluster,
SIG Scheduling, and SIG Storage as the stakeholder SIGs for the Policy
working group based on the working group's proposal email.

SIG On Premise was also listed in the email, but is not included in this
list as the SIG has been archived.

The Multitenancy and Resource Management working groups were also listed
in the email, but are not included in this list.