Create issue and PR management guildlines for the sig contribex
These guidelines and responsibilities are focused on  issues and PRs
that are identified for sig contributor experience. Keeping it as short
as possible for easier usage. It's also  utilizing and poiting to other
existing materials on triage.
Add the triage team members and specific role descriptions in a separate
document.

Related #3364