github: update instructions for creating repos from template repo
Using the current instructions, we can only create repos using the
template repo if the template repo is in the same org as the new repo.

If template repo is _not_ in the same org, we can still use it by visiting
the template repo page and clicking on "Use this template".