Add more details to generate-devstats-repo-sql.py
There was still a lot of manual massaging of the script's
output that needed to happen before it was usable as a PR
to cncf/devstats.

I have now tried to encode as much of that knowledge in
code and comments here to allow the output of the script
to overwrite the existing file.

- add header/footer
- add comment pointing to this script
- add special case group for "Kubernetes"
- add support for committees that own code
- (manually) keep track of old repo names for history