code of conduct: expand the list of people whose privacy will be respected
While we later said that any leak of personal information is inappropriete, this
sentence only specifically mentioned the right of privacy of the reporter. We believe
the reporter, anyone named in the report, anyone we talk to about the report, and
everyone involved should have their privacy respected. So this language is more
broad to make our intent more clear.