cherry-picks: bump up guidance on criteria, add table of contents
We see a relatively high volume of cherry picks that aren't clearly
meeting the criteria for cherry-picking, but that criteria has only
recently started to be in written, shared form and it is not super
discoverable.  To make it more clear let's make that its own section
(linkable), move it earlier in the document, and add a table of contents
to the cherry-picks.md for navigation.

Signed-off-by: Tim Pepper <tpepper@vmware.com>