Add Slack channel for Project Antrea
Antrea (https://github.com/vmware-tanzu/antrea) is an open-source CNI
plugin based on Open vSwitch. It was built specifically to be a
networking solution for Kubernetes clusters and supports Network
Policies.

We are adding a single channel at the moment (antrea) but may want to
have a dev channel in the future.

Signed-off-by: Antonin Bas <abas@vmware.com>