Fix formatting of OOM score configuration section
Two things were confusing to me in this section.
1. "The base OOM score is between 0 and 1000, so if process A's OOM_SCORE_ADJ - process B's OOM_SCORE_ADJ is over a 1000, then process A will always be OOM killed before B."
This implies that both A and B have the same base score? And could it be simplified to say if A's OOM_SCORE_ADJ > B, than A will be killed. And this would also apply to the final OOM score, if this is saying they both have the same base OOM score.
2. Pods, containers, and processes are mixed into a bunch of these rules and it's not super clear in each rule which one the OOM score applies to.
Ex.
- The final OOM score of a process is also between 0 and 1000. (process)
- If a burstable pod uses less memory than requested, then its OOM_SCORE < 1000 (pod)
- Pod infra containers or Special Pod init process (containers?)