Move cluster-proportional subprojects to sig-network
The OWNERS at the root of these repos are sig-network people, and
the recent activity I've seen around these repos is coming from
sig-network people. My understanding is sig-autoscaling doesn't
want these projects anymore.

This is the first step, the next step is to suggest that these be
moved to kubernetes-sigs or kubernetes-retired depending on your
willingness to support or own them.