Correct SIG Instrumentation meeting time
The meeting is at 9:30 Pacific Time.  As we are currently observing daylight savings, PDT = UTC-7, not UTC-8.  This off-by-one is observable by following the `Convert to your timezone` link.

If it is preferable, we could list the time as 15:30 UTC, but as UTC does not observe daylight savings, this would require an update every six months if we keep a Pacific-centric schedule.  Or, if we wish to adhere to a 16:30 UTC meeting, we should be meeting at 10:30 Pacific.

In the same vein, the weekly meeting notes list the meeting as starting at 9:30 PST.  We are currently in PDT.  This should be updated to simply read simply `PT` or `Pacific Time`.