resolved merge conflict
I don't think we need to use an entirely separate GOPATH, thus forcing
us to re-download modules every time we run `make`

This was causing verify-generated-docs to fail locally since go will
set its modcache as readonly and thus cause the cleanup "rm -rf" in
this script to fail. In go1.14 or later we could use "-modcacherw"
to stop making the modcache readonly but that bring me back to.. why
do need an entirely separate GOPATH in the first place?