Proposed charter for SIG Security (#4962)
* Proposed charter for SIG Security

Letter to Steering committee to be linked later.

* Update wg-security-audit/sig-security-charter-proposal.md

Co-authored-by: Micah Hausler <micahhausler@users.noreply.github.com>

* Accept nested list modification

Co-authored-by: Micah Hausler <micahhausler@users.noreply.github.com>

* Adding README created by community make.
Adding sig-security to sigs.yaml
Creating sig-security/ and moving proposed charter to that directory.

* Proposed charter for SIG Security

Letter to Steering committee to be linked later.

* Update wg-security-audit/sig-security-charter-proposal.md

Co-authored-by: Micah Hausler <micahhausler@users.noreply.github.com>

* Accept nested list modification

Co-authored-by: Micah Hausler <micahhausler@users.noreply.github.com>

* Correct capitalization

Co-authored-by: Nikhita Raghunath <nikitaraghunath@gmail.com>

* resolved merge conflict

I don't think we need to use an entirely separate GOPATH, thus forcing
us to re-download modules every time we run `make`

This was causing verify-generated-docs to fail locally since go will
set its modcache as readonly and thus cause the cleanup "rm -rf" in
this script to fail. In go1.14 or later we could use "-modcacherw"
to stop making the modcache readonly but that bring me back to.. why
do need an entirely separate GOPATH in the first place?

* Corrected employers

Co-authored-by: Micah Hausler <micahhausler@users.noreply.github.com>
Co-authored-by: Nikhita Raghunath <nikitaraghunath@gmail.com>