Add a slack channel for cluster-api-aws-alerts
This channel is intended to be used for routing of alerts that
affect tests running for the cluster-api-provider-aws subproject.

These alerts are currently routed to test-infra-oncall, and this will
allow for the alerts to be routed to the community members that are
more likely able to take action on the alerts.