Update channels.yaml to include new slack channel.
CRAFT is a project developed at Salesforce to facilitate easy operator creation. CRAFT declares Kubernetes Operators in a robust and generic way for any resource, letting developers focus on CRUD (create, read, update and delete) operations of resource management in a Dockerfile. With CRAFT you can create operators without a dependent layer and in the language of your choice. Further info: https://github.com/salesforce/craft#custom-resource-abstraction-fabrication-tool
We would like to create a channel in the Kubernetes community which can be used to raise and solve issues and suggest improvements on the existing framework and documentation.