Move sig-scale meetings to 10:30 PT
We decided that it'd be better than having the meeting time in UTC. Using UTC results in having the meeting at one time during half of the year and then at other time during another half. It makes hard as it requires re-organizing your calendar every 6 months. Using PT will keep the time the same for America. In Europe, we'll have two meetings at different time but we're fine with that.