SIG Cloud Provider 2020 Annual Report
Adding initial template for SIG Cloud Provider 2020 Annual Report.
Part of issues/5500.
This is due by March 1st and will be shown to the SIG and Slack.

Factored in first round of community feedback.
Tried to make sure we had answers for all questions.
Fix a few typos and a related owner issue.
Added nckturner to extraction owners.