Fix descriptions in sig-release usergroups.
Slack has a 140-character max for descriptions, and they can't contain
newlines. Tempelis doesn't understand either of these rules, so it is
confused: it is repeatedly trying to add in the newlines that Slack is
stripping out, and it cannot correctly create security-release-team because
the description is too long. Fix both of these things.