github-management: update org invite SLO to one week
The new membership coordinators also have low bandwidth and are actively
participating in leadership roles in other parts of the project.
We are also in a pandemic!

This commit updates the SLO to match the current state - we mostly take
upto a week to address org invites. We can switch it back to 72 hours
once folks have more time.