slack/sig-release: s/security-release-team/security-rel-team
When initially attempting to create this usergroup, its description
was too long for Slack to accept. The original group name exists...
somewhere... in Slack, but it's no longer possible to update it (despite
shortening the description in previous follow-ups).

Here we rename the group so it can be properly created by Tempelis.

Signed-off-by: Stephen Augustus <foo@auggie.dev>