request slack channel for Kilo
I'd like to request a slack channel for [Kilo](https://github.com/squat/kilo), a networking provider for Kubernetes built on WireGuard. This would be helpful to provide a designated place for the community members and maintainers to talk, discuss issues, and brainstorm.

As this is my first such request, please let me know if I need to do anything else.