New channel for security assessment of cluster api
From SIG-Security's last weekly meeting an action item was to create a new channel in order to facilitate a focussed discussion on security assessment of cluster-api (a sub project of kubernetes). This is a first attempt of a security assessment of a sub-project performed in collaboration with sig-security of the parent graduated project. 

Since we are breaking new ground here, I am happy to revisit if this channel should be a prefixed as `wg` or `sig`.  We intend to periodically revisit the results from our assessments, so I do not see this as a time-bound effort but more of a continuous improvement cycle.