communication/slack-config: archive headlamp-dev and rename -users
A while ago, the #headlamp channel was split into -users and -dev
because some members were concerned that an eventual dev conversation
increase would bother non-dev users.
However, our community is still small, and the Slack channels are
very low traffic, so this split only makes the communication more
difficult.

Therefore, as Headlamp's project manager, I would like to request
that the -users channel be renamed back to the prefixless case, and
the -dev channel to be archived.