github-management: add rotation program
This commit adds a rotation program for both GitHub Admins and NMCs
to create a sustainable culture and avoid burn out. This is also
important to ensure that requests from community members are handled
as per our SLO of one week.

Note that this is not an oncall program. Our SLO continues to remain
one week.