remove historic CoCC bootstrapping process document
This document has useful historical context, but may cause reader
confusion relative to the enacted state of the CoCC now that we are well
past boostrapping that committee.  Most, if not all, of the information
in this document is reflected in other documents which are curated for
accuracy.  It will remain in the git history for posterity.

Signed-off-by: Tim Pepper <tpepper@vmware.com>