contributors/devel/sig-api-machinery/controllers: Group workers for shutdown handling
Pivoting to the ConditionWithContextFunc pattern [1], so the various
handlers can say "don't bother launching me again, I'm satisfied" or
"yikes, this error is terrible, I'm giving up".  As an unlikely
example, perhaps something out of band shuts down the queue and
c.queue.Get() returns true for quit/shutdown, in which case the
controller will now respect [2]:

> If shutdown = true, the caller should end their goroutine.

A more likely scenario is that syncHandler, which is not defined in
the example, sees something that it considers not worth continuing on.
With this change, it will requeue with a rate limit, and return 'true,
nil', which will cause PollImmediateUntilWithContext to wrap things
up, and that particular worker goroutine will exit.  Any sibling
workers will pick up that key later, come to their own conclusions,
and possibly exit as well.

[1]: https://pkg.go.dev/k8s.io/apimachinery/pkg/util/wait#ConditionWithContextFunc
[2]: https://pkg.go.dev/k8s.io/client-go/util/workqueue#Type.Get