sig-cl: remove the kubespray Zoom meeting
As discussed with the kubespray maintainers:
https://kubernetes.slack.com/archives/C2V9WJSJD/p1643716538243509

Their Zoom meeting is no longer active.
The maintainers expressed that Slack is sufficient for them
and we should remove the meeting from the calendar for the time being.

Remove the meeting from sigs.yaml
It was also removed from the SIG CL calendar.