Add recommendation to use Informers to SLO guidance
If an app running in the cluster needs a constantly up-to-date
list of objects, and it does so by making repeated LIST calls,
that can add very significant load to etcd and the API server.
That is exactly the problem that Informers are designed to solve.

As of early 2022, the Haskel and Python libraries are the only
officially-supported libraries that do not yet include Informers.