Update channels.yaml
Adding a crossing channel #finsops to enable folks in the Kubernetes community to engage and participate with the FinOps Community.
Right now,the FinOps Slack is restricted to members who are certified. This channel is to allow the FInOps SIGs focused on Kubernetes(and containers) to have a broader conversation with the Kubernetes Community.