Make slack-guidelines links explicit
This doc is imported into the kubernetes.dev site to pull together
community docs from various locations. This works great for avoiding
duplication, but brings an implicit requirement that the links in the
doc can no longer be relative.

To avoid getting 404 links on kubernetes.dev, this updates all of the
links in `slack-guidelines.md` to be their full explicit path on GitHub.

Signed-off-by: Sean McGinnis <stmcg@amazon.com>