Add #podman-desktop to the Kubernetes slack
<!--  Thanks for sending a pull request!  Here are some tips for you:
- If this is your first contribution, read our Getting Started guide https://github.com/kubernetes/community/blob/master/contributors/guide/README.md
- If you are editing SIG information, please follow these instructions: https://git.k8s.io/community/generator
  You will need to follow these steps:
  1. Edit sigs.yaml with your change
  2. Generate docs with `make generate`. To build docs for one sig, run `make WHAT=sig-apps generate`
-->

**Which issue(s) this PR fixes**:
N/A

**Overview:**

Hey all!

We're a dev team that focuses on creating a GUI interface
for "Podman" which is an open source project.

Podman Desktop implements Kubernetes heavily as you can create a
container, change it to a pod, and then deploy to Kubernetes. All with
open source tools.

Under-the-hood it is podman (open source) as well as the kubernetes
Node.JS library.

We have first class support for Kubernetes and wish to
be part of the Kubernetes slack for all community communication.

Our future intentions of the project is the implementation of Docker
Compose to Kubernetes tools such as Kompose
(https://github.com/kubernetes/kompose), which I'm a core maintainer of.

We have a ROADMAP to add more Kubernetes features
(https://github.com/orgs/containers/projects/2), specifically: https://github.com/containers/podman-desktop/issues/29

**Purpose of the channel:**

Over the past year, we've been using Discord. Our motivation to join the
Kubernetes Slack is because all of us are already on it! We all already
have accounts and are active in other channels.

**Criteria:**

I Believe we fit all the critera as we're:

* An open source project
* Kubernetes-specific with first class support
* Not product focused or proprietary

**Chatroom requested:**

#podman-desktop

**About the team:**

* Charlie Drage (cdrage): Maintainer of http://github.com/kubernetes/kompose which already has a Kubernetes #kompose channel

* Florent Benoit (benoitf): Contributor and active maintainer in the
https://github.com/containers

* Vladyslav Zhukovskyi (vzhukovs): Contributor and active maintainer of
  https://github.com/containers

**Other:**

* We understand and acknowledge the Kubernetes Code of Conduct
* Hi from Kubecon (was there yesterday!)

Many thanks and let us know if you need anything else!

Signed-off-by: Charlie Drage <charlie@charliedrage.com>