Correct reference link formatting
When using reference links (`[a link to something][something]`) the link
goes itself is referred to later in the file. This normally should be a
line that starts with the reference name in brackets, followed by the
actual link (`[something]: <link>`).

Many of these reference links in the repo have unnecessary leading
spaces. While this renders OK, at least with GitHub's markdown
rendering, it currently causes problems for sites like kubernetes.dev
that pull in these docs and attempt to update the links so they are
usable from a different rendered location.

While the script that does that import should probably be updated to be
a little more robust, this addresses the immediate need of fixing these
broken kubernetes.dev links by removing the leading whitespaces so they
are properly recongized and updated.

Also some very minor cleanup in some older files. While these don't
appear to be imported elsewhere, they were cleaned up to avoid the risk
of any future copy/paste issues that could happen prior to the doc
import script being updated to handle things correctly.

Signed-off-by: Sean McGinnis <stmcg@amazon.com>