Remove community meeting requirements from governance docs
The community meeting has not been an effective means of communicating important information or updates for a few years now, and is no longer occurring on a regular cadence (and may be removed altogether soon in the future).

As such it should be removed as a community group governance requirement.