request to add kcp-edge-devs group
changes to ownership and strategy for kcp is requiring kcp-edge to change strategy and establish our own slack channel, mailing list, and etc.  We would like to continue working within the kubernetes ecosystem as we are expanding our support to work with a variety of different kubernetes distributions, not just kcp.