add eksctl slack channel
This is a request for the [eksctl](https://eksctl.io/) project to have a channel in the Kubernetes Slack.

eksctl is a simple CLI tool for creating and managing clusters on EKS - AWS's managed Kubernetes service. The project is open source, and we want to move our Slack channel to Kubernetes Slack.