request to add kubestellar-devs group (#7275)
* request to add kcp-edge-devs group

changes to ownership and strategy for kcp is requiring kcp-edge to change strategy and establish our own slack channel, mailing list, and etc.  We would like to continue working within the kubernetes ecosystem as we are expanding our support to work with a variety of different kubernetes distributions, not just kcp.

* Update usergroups.yaml

corrected indent on line 294

* updated user names to match GH

* Update usergroups.yaml

name of project has changed

* Update communication/slack-config/usergroups.yaml

Co-authored-by: Madhav Jivrajani <12381034+MadhavJivrajani@users.noreply.github.com>

* Update usergroups.yaml

converted to lowercase

* Update usergroups.yaml

I had names from PROW in here - they are now kubernetes slack org IDs.  Sorry about that.

---------

Co-authored-by: Madhav Jivrajani <12381034+MadhavJivrajani@users.noreply.github.com>